+++ web/src/components/Navigation/Background/Background.component.tsx
@@ -1,3 +1,5 @@
+import './Background.css'
+
 const Background = () => {
     return (
         <div className="gradient-background fixed top-0 block w-screen h-screen z-[-999]">
