+++ web/src/components/Navigation/NavBar/NavBar.css
@@ -1,0 +1,3 @@
+.navbar-inner-shadow {
+    box-shadow: inset -8px 0px 18px rgba(0, 0, 0, 0.15);
+}
